requirements.txt
flask==3.0.0
flask-cors==4.0.0
transformers==4.36.0
torch==2.1.0
PyPDF2==3.0.1
python-docx==1.0.0
reportlab==4.0.7
numpy==1.24.3
tqdm==4.66.1
regex==2023.10.3
requests==2.31.0 